Video Player v2/requirements.txt
opencv_python_headless==4.5.4.60
Pillow==9.5.0